+++ src/components/links/link/Link.svelte
@@ -4,19 +4,35 @@
     import Icon from "../../../icons/Icon.svelte";
     import { type IconName } from "../../../icons/types/IconName";
     import { type IconLocation, iconLocationDefault } from "./LinkIconLocation";
+    import openInNewWindow from "./utils/openInNewWindow";
+
+    const onClick = (event: MouseEvent) => {
+        if (href !== "" && opensNewWindow === "true") {
+            event.preventDefault();
+            event.stopPropagation();
+            openInNewWindow(href);   
+        }
+    }
 
    
     export let title: string = "";
-
-    export let href: string | null = null;
+    export let href: string = "";
     export let rel: string | null = null;
     export let target: string | null = null;
     export let iconName: IconName | "" = "";
     export let iconLocation: IconLocation = iconLocationDefault;
+    export let opensNewWindow: string = "false";
 
 </script>
 
-<a class="link {iconName !== "" ? (iconLocation === 'after') ? 'link-with-icon-after' : 'link-with-icon-before': ''}" href={href} title={title} rel={rel} target={target}>
+<a 
+    class="link {iconName !== "" ? (iconLocation === 'after') ? 'link-with-icon-after' : 'link-with-icon-before': ''}" 
+    href={href} 
+    title={title} 
+    rel={rel} 
+    target={target}
+    on:click={onClick}
+    >
     <slot>Empty Link</slot>
     {#if iconName !== ""}
     <span class="icon" aria-hidden="true">
@@ -40,6 +56,7 @@
         text-underline-offset: var(--khao-sys-size-regular-1);
         cursor: pointer;
         vertical-align: baseline;
+        margin-right: var(--khao-link-space-to-next-char);
     }
 
     .link-with-icon-after {
@@ -47,7 +64,6 @@
         flex-direction: row;
         align-items: center;
         gap: var(--khao-link-icon-space);
-        margin-right: var(--khao-link-space-to-next-char);
     }
 
     .link-with-icon-before {
@@ -55,6 +71,7 @@
         flex-direction: row-reverse;
         align-items: center;
         gap: var(--khao-link-icon-space);
+        margin-right: 0;
         margin-left: var(--khao-link-space-to-next-char);
     }