+++ src/index.ts
@@ -5,7 +5,6 @@ export * as Button from "./components/buttons/button/Button.svelte";
 export * as Card from "./components/cards/card/Card.svelte";
 export * as MarkerPin from "./components/map/markerPin/MarkerPin.svelte";
 export * as Link from "./components/links/link/Link.svelte";
-export * as NoticeCard from "./components/cards/noticeCard/NoticeCard.svelte";
 export * as PayPalDonateButton from "./components/buttons/payPalDonateButton/PayPalDonateButton.svelte";
 export * as ShareButtons from "./components/widgets/shareButtons/ShareButtons.svelte";
 export * as SteadyButton from "./components/buttons/steadyButton/SteadyButton.svelte";