+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [1.3.0]
 
+- added `newsletterTeaser` attribute to `khao-social-buttons`
 - added icons `navigate-back` and `navigate-forward` to to `khao-icon`
 - added `search` icon to `khao-icon`
 - changed help icon in `khao-icon`