+++ CHANGELOG.md
@@ -5,13 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0]
+
+- added `newsletterTeaser` attribute to `khao-social-buttons`
+- added icons `navigate-back` and `navigate-forward` to to `khao-icon`
+- added `search` icon to `khao-icon`
+- changed help icon in `khao-icon`
+- added `opensNewWindow` attribute to `khao-link`
+- (!) ommited `khao-card-notice`, you can use `khao-card` to layout card accordingly
+
+## [1.2.3]
+
+- added icons `arrow-down`, `arrow-left`, `arrow-right`, and `arrow-up`
+- added icon `play`
+- added aria-label to `khao-card`
+
+## [1.2.2]
+
+- deprecated `khao-card-notice`
+- improved layout for `khao-link`
+
 ## [1.2.1]
 
-- fixed default `iconLocation` behavior
+- fixed default `iconLocation` behavior on `khao-link`
 
 ## [1.2.0]
 
-- !changed `khao-link` interface: label now has to given as child element.
+- (!) changed `khao-link` interface: label now has to given as child element.
 - improved `khao-link` display
 - added `iconLocation` prop to `khao-link`